% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arl_reader.R
\name{ARL_read}
\alias{ARL_read}
\title{Read an ARL file}
\usage{
ARL_read(file_to_read, var_i_want, ll, ur, latlon = TRUE, verbose = FALSE)
}
\arguments{
\item{file_to_read}{the name of the file to read in}

\item{var_i_want}{the name of the variable to extract}

\item{ll}{the lat and lon of the lower-left corner of the area to extract}

\item{ur}{the lat and lon of the upper-right corner of the area to extract}

\item{latlon}{whether to convert the raster to lat-lon grid. Otherwise it stays in the native projection}

\item{verbose}{whether to output additional messages during extraction'}
}
\value{
for 2d variables, a raster time series object is returned. For 3d variables, a list of RTS objects is returned, one for each altitude.
}
\description{
Read an ARL file
}
\examples{
file_to_read <- 'hysplit.20161020.00z.hrrra'
lower_left <- c(-73,40)
upper_left <- c(-70,43)
pbl <- ARL_read(file_to_read,"PBLH",ll=lower_left,ur=upper_right)
plot(pbl)
}
